prequals/UnderContruction/deploy/src/requirements.txt
flask
py3-validate-email
